quickdraw-doodle-recognition/requirements.txt
absl-py==0.5.0
astor==0.7.1
astroid==2.0.4
backcall==0.1.0
bitarray==0.8.3
bleach==3.0.1
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.2.3
flake8==3.5.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kaggle==1.5.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lightfm==1.15
llvmlite==0.25.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.3
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numba==0.40.1
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.1
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
py-cpuinfo==4.0.0
py4j==0.10.7
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==2.1.1
pymongo==3.7.2
pyparsing==2.2.2
pyspark==2.3.2
pytest-runner==4.2
python-dateutil==2.7.3
python-slugify==1.2.6
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.1
rdp==0.8
requests==2.20.0
rope==0.11.0
sacred==0.7.4
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplification==0.3.8
six==1.11.0
sklearn==0.0
snakeviz==1.0.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thrift==0.11.0
tornado==5.1.1
tqdm==4.27.0
traitlets==4.3.2
typed-ast==1.1.0
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd==1.1.0
